Promote enemy attribute build to $lib/common; delegate enemyDefense
Generalises the Codex's enemyAttributesAtLevel into a single, store-free
$lib/common module (enemy-attributes.ts) that owns the enemy-at-level battle
build. A new memoised enemyBattleAttributes is the single source both display
surfaces consume: enemyAttributesAtLevel derives the dossier's primary/secondary
values from it, and the Skills screen's enemyDefense now delegates to it instead
of duplicating the BattleAttributes composition and its own WeakMap cache.

- Move per-enemy/level WeakMap memoisation into the promoted module.
- Update codex-view and skills-view imports; drop the duplicated cache + the
  now-unused IEnemy import in skills-view.
- Relocate the unit tests to tests/lib/common, add enemyDefense coverage and a
  delegation/parity test against enemyAttributesAtLevel; remove the old module
  and tests.
